Refactor review modal and Application Table
